♻️Refactor: ChartTimeScaleType 전역 변수로 변경 #61
