+++ view/src/shared/types/ChartTimeScaleType.ts
